+++ docs/source/changelog.rst
@@ -3964,3 +3964,7 @@ Current C++/SQLite client, Python/SQLAlchemy server
 
 - Update to use SQLAlchemy 2.0 (Database revision 0086 for minor changes to
   comments for HAMD and HAMD7).
+  https://github.com/ucam-department-of-psychiatry/camcops/issues/322
+
+- Qt version now 6.5.5. This fixes the display of the About Qt dialog on iOS
+  https://github.com/ucam-department-of-psychiatry/camcops/issues/308